// source/types/training-type.ts
// The existing TrainingType members order is needed for the legend to be displayed properly
export enum TrainingType {
  speed = 'speed training',
  strength = 'strength training',
  race = 'race',
  timed = 'timed training',
  moderate = 'moderate training',
  recovery = 'recovery training',
  rest = 'rest'
}